+++ example/input.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:commander_ui/commander_ui.dart';
 
 final class Item {
@@ -13,20 +15,20 @@ Future<void> main() async {
   final input = Input(
       answer: 'Please give us your name',
       placeholder: 'firstname lastname',
-      validate: (value) =>
-      switch(value) {
-        String value when value
-            .trim()
-            .isNotEmpty => Ok(null),
-        _ => Err('Please provide a valid name')
-      }
-  );
+      secure:
+          true, // 👈 Optional, ou can hide the input like html input with password type
+      validate: (value) => switch (value) {
+            String value when value.trim().isNotEmpty => Ok(null),
+            _ => Err('Please provide a valid name')
+          });
 
-  final value = switch(await input.handle()) {
+  final value = switch (await input.handle()) {
     Ok(:final value) => 'My value is $value',
     Err(:final error) => Exception('Error: $error'),
     _ => 'Unknown',
   };
 
   print(value);
+
+  exit(0);
 }