+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Implement password secure
+- Restore cursor after component result
+
 ## 1.0.0
 
 - Initial version.